Extracted TargetValidator
